Fix race condition on `wait()`
`wait()` using `oneshot::channel()` makes race conditions. Call
`Executor::wait()` directly.

It is possible that the thread that has executed `wait()`, which is
usually the main thread, terminates before the thread executing
`done_tx.send(())` drops `rpc_handler`.
Static variables are destructed at the end of termination of the main
thread, and then a segfault occurs when the thread dropping the
`rpc_handler` accesses the variables.