Fix race condition on `wait()` (#504)
Manually dropping the future passed from `Future::select` before sending
`done_tx` prevents the race condition.

`Future::select` pass the unresolved future, which is a `Server` holding
`rpc_handler`, to the following callback. Therefore, it is dropped after
the `done_tx.send(())` after the callback exits.
It is possible that the thread that has executed `wait()`, which is
usually the main thread, terminates before the thread sending `done_tx`
drops `rpc_handler`.
Static variables are destructed at the end of termination of the main
thread, and then a segfault occurs when the thread dropping the
rpc_handler accesses the variables.